refactor: move modal into components (prepare a mode autonomous components)
