+++ static/js/features/library/music.js
@@ -3,7 +3,7 @@ import { getCsrfHeaders } from '../../core/csrf.js';
 import { formatFileSize } from '../../core/formatters.js';
 import { i18n } from '../../core/i18n.js';
 import { oxiIcon } from '../../core/icons.js';
-import { Modal } from '../../core/modal.js';
+import { Modal } from '../../components/modal.js';
 import { notifications } from '../../core/notifications.js';
 
 /** @import {FileItem, Musicshare, Playlist, PlaylistItem} from '../../core/types.js' */
