+++ static/js/workers/README.md
@@ -1,0 +1,1 @@
+Put here all workers you don't want to be bundled into tha main application js
